% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils.R
\name{app_sys}
\alias{app_sys}
\title{Locate a file installed under inst/}
\usage{
app_sys(...)
}
\arguments{
\item{...}{Path components relative to inst/.}
}
\value{
The full path returned by \code{\link[=system.file]{system.file()}} for this package.
}
\description{
Locate a file installed under inst/
}
