% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_detect.R
\name{autocrop_board}
\alias{autocrop_board}
\title{Best-effort crop to the board by trimming near-uniform margins}
\usage{
autocrop_board(img)
}
\arguments{
\item{img}{A magick image of a board, possibly with surrounding margins.}
}
\value{
A magick image cropped to the detected square board region.
}
\description{
Best-effort crop to the board by trimming near-uniform margins
}
