% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_piecesets.R
\name{available_piece_sets}
\alias{available_piece_sets}
\title{Names of all piece sets available locally}
\usage{
available_piece_sets()
}
\value{
A character vector of set names (bundled cburnett plus any
downloaded manifest sets).
}
\description{
Names of all piece sets available locally
}
