% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils.R
\name{custom_templates_path}
\alias{custom_templates_path}
\title{Path for user-calibrated ("custom") templates}
\usage{
custom_templates_path()
}
\value{
The RDS path where a manual calibration is persisted across
restarts.
}
\description{
Path for user-calibrated ("custom") templates
}
