% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_theme.R
\name{cv_theme}
\alias{cv_theme}
\title{The application's Bootstrap theme}
\usage{
cv_theme()
}
\value{
A \code{\link[bslib:bs_theme]{bslib::bs_theme()}} object.
}
\description{
The application's Bootstrap theme
}
