% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils.R
\name{decode_data_url_image}
\alias{decode_data_url_image}
\title{Decode a data-URL image string into a magick image}
\usage{
decode_data_url_image(data_url)
}
\arguments{
\item{data_url}{A "data:image/...;base64,...." string.}
}
\value{
A magick image.
}
\description{
Decode a data-URL image string into a magick image
}
