% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_engine.R
\name{ensure_stockfish}
\alias{ensure_stockfish}
\title{Return a path to a Stockfish binary, downloading it on first use}
\usage{
ensure_stockfish()
}
\value{
The path to a usable Stockfish binary.
}
\description{
Return a path to a Stockfish binary, downloading it on first use
}
