% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rating_estimate.R
\name{fen_turn}
\alias{fen_turn}
\title{Which side is to move in a FEN}
\usage{
fen_turn(fen)
}
\arguments{
\item{fen}{A FEN string.}
}
\value{
"w" or "b".
}
\description{
Which side is to move in a FEN
}
