% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_engine.R
\name{find_local_engine}
\alias{find_local_engine}
\title{Find a previously-downloaded Stockfish binary in the cache}
\usage{
find_local_engine()
}
\value{
The path to the binary, or \code{NULL} if none is present.
}
\description{
Find a previously-downloaded Stockfish binary in the cache
}
