% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_game_tracker.R
\name{game_current_fen}
\alias{game_current_fen}
\title{The position a tracked game currently stands at}
\usage{
game_current_fen(game)
}
\arguments{
\item{game}{A game record from \code{\link[=game_new]{game_new()}}.}
}
\value{
The latest FEN.
}
\description{
The position a tracked game currently stands at
}
