% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_game_tracker.R
\name{game_ply}
\alias{game_ply}
\title{Number of plies played in a tracked game}
\usage{
game_ply(game)
}
\arguments{
\item{game}{A game record from \code{\link[=game_new]{game_new()}}.}
}
\value{
The move count.
}
\description{
Number of plies played in a tracked game
}
