% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils.R
\name{golem_add_external_resources}
\alias{golem_add_external_resources}
\title{Register www/ as a resource path and add css/js includes}
\usage{
golem_add_external_resources()
}
\value{
A \code{tags$head} include with the stylesheet and paste script.
}
\description{
Register www/ as a resource path and add css/js includes
}
