% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_chessjs.R
\name{is_valid_fen}
\alias{is_valid_fen}
\title{Test whether a FEN is legal}
\usage{
is_valid_fen(ctx, fen)
}
\arguments{
\item{ctx}{A chess.js V8 context from \code{\link[=new_chess_context]{new_chess_context()}}.}

\item{fen}{A FEN string.}
}
\value{
\code{TRUE} if the FEN is legal, \code{FALSE} otherwise.
}
\description{
Test whether a FEN is legal
}
