% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rating_estimate.R
\name{maia_pool_stop}
\alias{maia_pool_stop}
\title{Stop a pool of Maia sessions}
\usage{
maia_pool_stop(pool)
}
\arguments{
\item{pool}{A pool from \code{\link[=maia_pool_start]{maia_pool_start()}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Stop a pool of Maia sessions
}
