% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_human_model.R
\name{maia_session_stop}
\alias{maia_session_stop}
\title{Stop a Maia session}
\usage{
maia_session_stop(sess)
}
\arguments{
\item{sess}{A session from \code{\link[=maia_session_start]{maia_session_start()}}.}
}
\value{
Invisibly \code{NULL}.
}
\description{
Stop a Maia session
}
