% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analyze.R
\name{mod_analyze_ui}
\alias{mod_analyze_ui}
\title{Analyze tab UI}
\usage{
mod_analyze_ui(id)
}
\arguments{
\item{id}{The module id.}
}
\value{
A Shiny UI tag list for the analyze tab.
}
\description{
Analyze tab UI
}
