% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calibrate.R
\name{mod_calibrate_ui}
\alias{mod_calibrate_ui}
\title{Calibrate tab UI}
\usage{
mod_calibrate_ui(id)
}
\arguments{
\item{id}{The module id.}
}
\value{
A Shiny UI tag list for the piece-sets and calibration tab.
}
\description{
Calibrate tab UI
}
