% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_chessjs.R
\name{new_chess_context}
\alias{new_chess_context}
\title{Create a chess.js V8 context}
\usage{
new_chess_context()
}
\value{
A V8 context with \code{ChessCtor} and \code{validateFenJS} bound, ready for
\code{\link[=validate_fen]{validate_fen()}} and \code{\link[=fen_move_to_san]{fen_move_to_san()}}.
}
\description{
Create a chess.js V8 context
}
