% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_piecesets.R
\name{piece_set_ext}
\alias{piece_set_ext}
\title{Find which file extension a downloaded piece set uses}
\usage{
piece_set_ext(dir)
}
\arguments{
\item{dir}{A piece-set directory.}
}
\value{
The extension for which all 12 piece files are present, or \code{NULL}.
}
\description{
Find which file extension a downloaded piece set uses
}
