% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the tanmai Shiny application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[golem:with_golem_options]{golem::with_golem_options()}}.}
}
\value{
Called for its side effect; runs the Shiny app.
}
\description{
Run the tanmai Shiny application
}
