% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_detect.R
\name{split_board}
\alias{split_board}
\title{Split a normalized board into 64 square images}
\usage{
split_board(board)
}
\arguments{
\item{board}{A magick image sized \code{BOARD_PX} by \code{BOARD_PX}.}
}
\value{
A list of 64 magick images, row-major, top-left square first.
}
\description{
Split a normalized board into 64 square images
}
