% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_recognize.R
\name{sq_core}
\alias{sq_core}
\title{Background-subtracted central region of a square}
\usage{
sq_core(mat)
}
\arguments{
\item{mat}{A \code{SQUARE_PX} by \code{SQUARE_PX} grayscale square matrix.}
}
\value{
The inner region with the estimated background subtracted.
}
\description{
Background-subtracted central region of a square
}
