% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_recognize.R
\name{sq_energy}
\alias{sq_energy}
\title{Mean absolute energy of a square core}
\usage{
sq_energy(core)
}
\arguments{
\item{core}{A background-subtracted square core from \code{\link[=sq_core]{sq_core()}}.}
}
\value{
The mean absolute value (used to decide empty vs occupied).
}
\description{
Mean absolute energy of a square core
}
