% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanmai-package.R
\docType{package}
\name{tanmai-package}
\alias{tanmai}
\alias{tanmai-package}
\title{tanmai: Chess Board Screenshot to Best Move}
\description{
A golem Shiny app that reads a chess board screenshot (drag-drop or clipboard
paste), recognizes the position with template matching against multiple piece
sets, and suggests the best move via a local Stockfish engine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tenmeh.github.io/tanmai/}
  \item \url{https://github.com/tenmeh/tanmai}
  \item Report bugs at \url{https://github.com/tenmeh/tanmai/issues}
}

}
\author{
\strong{Maintainer}: tenmeh \email{81738153+tenmeh@users.noreply.github.com}

Authors:
\itemize{
  \item tenmeh \email{81738153+tenmeh@users.noreply.github.com}
}

}
\keyword{internal}
