% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_detect.R
\name{to_gray_matrix}
\alias{to_gray_matrix}
\title{Convert a magick image to a grayscale integer matrix}
\usage{
to_gray_matrix(img)
}
\arguments{
\item{img}{A magick image.}
}
\value{
An integer matrix indexed \verb{[x, y]} (width by height), values 0-255.
}
\description{
Convert a magick image to a grayscale integer matrix
}
