% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analyze.R
\name{mod_analyze_server}
\alias{mod_analyze_server}
\title{Analyze tab server}
\usage{
mod_analyze_server(id, chess_ctx)
}
\arguments{
\item{id}{The module id.}

\item{chess_ctx}{A chess.js V8 context from \code{\link[=new_chess_context]{new_chess_context()}}.}
}
\value{
Called for its side effects; wires up the analyze reactives.
}
\description{
Analyze tab server
}
