% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_board.R
\name{mod_board_ui}
\alias{mod_board_ui}
\title{Interactive board UI}
\usage{
mod_board_ui(id)
}
\arguments{
\item{id}{The module id.}
}
\value{
A Shiny UI tag list for the interactive analysis board.
}
\description{
Interactive board UI
}
